requirements.txt
langchain
langchain_community
gpt4all
pydantic
hydra-core==1.3
unstructured[pdf]
langchain-chroma>=0.1.2
chromadb